Lots of chao machine hacking
So big changes here. Chao machine is semi-restored to how it looks on GC (this can be a whole separate mod). I've instead went and made my own dialog boxes and integration with the network class through the GC menus (these things are called Odekake and Toridasu in the debug symbols which mean drop off and pick up so I just carried on with the naming). For this commit, Odekake is integrated with the chao app to be able to send the main chao data if the connection is correct.